// vite-app/app/src/redux/interfaces.ts
export interface ProfileUserdata {
    organization: string | null;
    department: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    affiliations: Array<Affiliation>;
    researchStatement: string;
    jobTitle: string;
    jobTitleOther: string;
    researchInterests: Array<string>;
    researchInterestsOther: string | null;
    fundingSource: string;
    gravatarDefault: string;
    avatarOption: string;
}

export interface Affiliation {
    title: string;
    organization: string;
    started: number;
    ended: number | null;
}